[Feat] [FE-16] Clipboard 추가
